fix: rename dirs to reflect charm names and fix lib publish issue
