+++ .github/workflows/ci.yaml
@@ -12,13 +12,13 @@ on:
     types: [opened, synchronize, reopened, ready_for_review]
 
 jobs:
-  maas-agent-charm:
+  maas-agent:
     name: MAAS Agent unit test
     runs-on: ubuntu-latest
     timeout-minutes: 5
     defaults:
       run:
-        working-directory: ./maas-agent-charm
+        working-directory: ./maas-agent
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -32,13 +32,13 @@ jobs:
       - name: Run tests
         run: tox run -e unit
 
-  maas-region-charm:
+  maas-region:
     name: MAAS region unit test
     runs-on: ubuntu-latest
     timeout-minutes: 5
     defaults:
       run:
-        working-directory: ./maas-region-charm
+        working-directory: ./maas-region
     steps:
       - name: Checkout
         uses: actions/checkout@v4
