+++ maas-agent/src/charm.py
@@ -9,7 +9,7 @@ import socket
 
 import ops
 from charms.grafana_agent.v0 import cos_agent
-from charms.maas_region_charm.v0 import maas
+from charms.maas_region.v0 import maas
 from charms.operator_libs_linux.v2.snap import SnapError
 from helper import MaasHelper
