+++ maas-agent/tests/unit/test_charm.py
@@ -10,7 +10,7 @@ from unittest.mock import patch
 import ops
 import ops.testing
 from charm import MAAS_RACK_PORTS, MAAS_SNAP_CHANNEL, MaasRackCharm
-from charms.maas_region_charm.v0 import maas
+from charms.maas_region.v0 import maas
 
 
 class TestCharm(unittest.TestCase):
