+++ maas-agent/tox.ini
@@ -10,7 +10,7 @@ min_version = 4.0.0
 [vars]
 src_path = {tox_root}/src
 tests_path = {tox_root}/tests
-lib_path = {tox_root}/lib/charms/maas_region_charm/v0/
+lib_path = {tox_root}/lib/charms/maas_region/v0/
 all_path = {[vars]src_path} {[vars]tests_path} {[vars]lib_path}
 
 [testenv]
