+++ maas-region/src/charm.py
@@ -14,7 +14,7 @@ import ops
 import yaml
 from charms.data_platform_libs.v0 import data_interfaces as db
 from charms.grafana_agent.v0 import cos_agent
-from charms.maas_region_charm.v0 import maas
+from charms.maas_region.v0 import maas
 from helper import MaasHelper
 
 logger = logging.getLogger(__name__)
